+++ lightsim2grid/lightSimBackend.py
@@ -70,7 +70,20 @@ class LightSimBackend(Backend):
     if not hasattr(Backend, "detachment_is_allowed"):
         # for legacy grid2op (< 1.11.0)
         detachment_is_allowed = DEFAULT_ALLOW_DETACHMENT
-        
+    
+    KEYS_PYPOWSYBL_LOADER = {
+        "gen_slack_id",
+        "use_buses_for_sub",
+        "sort_index",
+        "init_vm_pu",
+        "sn_mva"
+    }
+    
+    KEYS_PANDAPOWER_LOADER = {
+        "pp_orig_file"
+    }
+    
+    
     def __init__(self,
                  detailed_infos_for_cascading_failures: bool=False,
                  can_be_copied: bool=True,
@@ -780,10 +793,11 @@ class LightSimBackend(Backend):
         loader_kwargs = {}
         if self._loader_kwargs is not None:
             loader_kwargs = self._loader_kwargs
-        
+        self._aux_check_loader_kwargs(loader_kwargs, "pypowsybl", type(self).KEYS_PYPOWSYBL_LOADER)
+                
         try:
             full_path = self.make_complete_path(path, filename)
-        except AttributeError as exc_:
+        except AttributeError as _:
             warnings.warn("Please upgrade your grid2op version")
             full_path = self._should_not_have_to_do_this(path, filename)
             
@@ -800,16 +814,30 @@ class LightSimBackend(Backend):
             buses_for_sub = False
             self.n_sub = df.shape[0]
             self.name_sub = ["sub_{}".format(i) for i, _ in enumerate(df.iterrows())]
+        
+        sort_index = True
+        if "sort_index" in loader_kwargs and (bool(loader_kwargs["sort_index"]) == loader_kwargs["sort_index"]):
+            sort_index = bool(loader_kwargs["sort_index"])
+        init_vm_pu = 1.06
+        if "init_vm_pu" in loader_kwargs and float(loader_kwargs["init_vm_pu"]) != loader_kwargs["init_vm_pu"]:
+            init_vm_pu = float(loader_kwargs["init_vm_pu"])
+        sn_mva = 1.06
+        if "sn_mva" in loader_kwargs and float(loader_kwargs["sn_mva"]) != loader_kwargs["sn_mva"]:
+            sn_mva = float(loader_kwargs["sn_mva"])
             
         n_busbar_per_sub = self._aux_get_substation_handling_from_loader_kwargs(loader_kwargs)
         if n_busbar_per_sub is None:
             n_busbar_per_sub = self.n_busbar_per_sub
         self._grid, subs_id = init_pypow(grid_tmp,
                                          gen_slack_id=gen_slack_id,
-                                         sort_index=True,
+                                         slack_bus_id=None,
+                                         sn_mva=sn_mva,
+                                         sort_index=sort_index,
+                                         only_main_component=self._automatically_disconnect,
                                          return_sub_id=True,
                                          n_busbar_per_sub=n_busbar_per_sub,
                                          buses_for_sub=buses_for_sub,
+                                         init_vm_pu=init_vm_pu,
                                          )
         (buses_sub_id, gen_sub, load_sub, (lor_sub, tor_sub), (lex_sub, tex_sub), 
          batt_sub, sh_sub, hvdc_sub_from_id, hvdc_sub_to_id) = subs_id
@@ -1004,11 +1032,33 @@ class LightSimBackend(Backend):
         self.init_pp_backend.load_grid(path, filename)
         self._aux_init_pandapower()
     
+    def _aux_check_loader_kwargs(
+        self,
+        loader_kwargs,
+        loader_name,
+        allowed_loader_kwargs):
+        for el in loader_kwargs:
+            if el not in allowed_loader_kwargs:
+                raise RuntimeError(f"Invalid 'loader_kwargs' {el} provided when loading a "
+                                   f"LightSimBackend with {loader_name} grids. Supported keys are: "
+                                   f"{sorted(allowed_loader_kwargs)}.")
+        
     def _aux_init_pandapower(self):
         from lightsim2grid.gridmodel import init_from_pandapower
+        pp_orig_file = "pandapower_v2"
+        loader_kwargs = {}
+        if self._loader_kwargs is not None:
+            loader_kwargs = self._loader_kwargs
+            
+        self._aux_check_loader_kwargs(loader_kwargs, "pandapower", type(self).KEYS_PANDAPOWER_LOADER)
+                
+        if "pp_orig_file" in loader_kwargs and str(loader_kwargs["pp_orig_file"]) == loader_kwargs["pp_orig_file"]:
+            pp_orig_file = str(loader_kwargs["pp_orig_file"])
+            
         self._grid = init_from_pandapower(self.init_pp_backend._grid,
                                           self.init_pp_backend.n_sub,
-                                          self.n_busbar_per_sub)    
+                                          self.n_busbar_per_sub,
+                                          pp_orig_file=pp_orig_file)    
         self.__nb_bus_before = self.init_pp_backend.get_nb_active_bus()  
         self._aux_setup_right_after_grid_init()