+++ lightsim2grid/tests/test_GridModel_pypowsybl.py
@@ -6,7 +6,6 @@
 # SPDX-License-Identifier: MPL-2.0
 # This file is part of LightSim2grid, LightSim2grid implements a c++ backend targeting the Grid2Op platform.
 
-from re import M
 import unittest
 import numpy as np